reorganized everything so its way nicer
