+++ .env.template
@@ -1,1 +1,1 @@
-WEATHER_API_KEY=
+WEATHER_API_KEY=weatherbit.io api key
