+++ src/components/WeatherTable.svelte
